traffic-backend/requirements.txt
# TrafficIQ Backend Dependencies

# Web Framework
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.0.0

# Configuration
python-dotenv>=1.0.0

# AI/ML - Vehicle Detection
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
torch>=1.9.0

# Production Server
gunicorn>=21.0.0
gevent>=23.0.0
gevent-websocket>=0.10.1

# Utilities
Pillow>=9.0.0

# Database
pymongo>=4.6.0
